requirements.txt
telethon
python-decouple
requests